format database

+++ src/TinyJobApi/Models/Vo/SchedulerUpdateVo.cs
@@ -7,7 +7,7 @@ namespace TinyJobApi.Models.Vo
         public string? Name { get; set; }
         public string? Description { get; set; }
         public SchedulerType? Type { get; set; }
-        public long?  PackageId { get; set; }
+        public int?  PackageId { get; set; }
         public string? AssemblyName { get; set; }
         public string? Namespace { get; set; }
         public string? ClassName { get; set; }
